refactor: move layout into ui module
