+++ rigflow-client/src/app/mod.rs
@@ -1,6 +1,6 @@
 pub mod bands;
 pub mod frequency_view;
-pub mod layout;
+//pub mod layout;
 pub mod om_bands;
 //pub mod state;
 pub mod stats;
