+++ rigflow-client/src/ui/mod.rs
@@ -1,7 +1,7 @@
 pub mod app;
 //pub mod bands;
 //pub mod frequency_view;
-//pub mod layout;
+pub mod layout;
 //pub mod om_bands;
 //pub mod spectrum_utils;
 pub mod state;
